recipes-bsp: Update loaders for Renesas BSP v5.9.0
To be precise: to the revisions which corresponds to the latest
commit of meta-renesas' "dunfell-dev" branch:
39334aeb54b167da09e108e9a654f852da48da68
"rcar-gen3: optee_os: Update SRCREV to Rev3.0.5"

Also maybe the settings of "RCAR_BL33_EXECUTION_EL=1" should be removed
from meta-xt-rcar-driver-domain/.../arm-trusted-firmware_git.bbappend
when building with AGL, but definitely shouldn't when building without
it.

Tested on salvator-x-h3-4x2g and h3ulcb-4x2g-kf machines.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>