misc: Update other components for Renesas BSP v5.9.0
To be precise: to the revisions which corresponds to the latest
commit of meta-renesas' "dunfell-dev" branch:
39334aeb54b167da09e108e9a654f852da48da68
"rcar-gen3: optee_os: Update SRCREV to Rev3.0.5"

Please note, we remove patch for U-boot because it didn't apply
anymore. It is harmless because it doesn't affect boot, so could be
updated afterwards.

Tested on salvator-x-h3-4x2g and h3ulcb-4x2g-kf machines.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>