recipes-graphics: Update for Weston 10 found in Kirkstone
With the current configuration, the 'XDG_RUNTIME_DIR' environment
variable is set only after login to the DomD. That is wrong and
causing the DomU initialization failure. The 'XDG_RUNTIME_DIR'
environment variable should be set right on the system's first boot
before the dependent services ( weston, displbe ) are started inside
the DomD.

This patch tries to address the wrong behavior by introducing the new
'preinit-weston' service.

Signed-off-by: Vladyslav Goncharuk vladyslav_goncharuk@epam.com